# specs/todo-fullstack/contracts/api-contract.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: REST API for Todo Full-Stack Web Application - Phase II
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Development server

paths:
  /{user_id}/tasks:
    get:
      summary: List all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '404':
          description: User not found

    post:
      summary: Create a new task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '422':
          description: Validation error

  /{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '404':
          description: Task not found

    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '404':
          description: Task not found

    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '204':
          description: Task deleted
        '401':
          description: Unauthorized
        '404':
          description: Task not found

  /{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
        user_id:
          type: string
          description: ID of the user who owns the task
        title:
          type: string
          description: Title of the task
          minLength: 1
          maxLength: 200
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000
        completed:
          type: boolean
          description: Completion status of the task
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the task
          minLength: 1
          maxLength: 200
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: Title of the task
          minLength: 1
          maxLength: 200
        description:
          type: string
          description: Detailed description of the task
          maxLength: 1000